Revert r4546:4548 (the mass removal of dynamic_cast<>).
We have always used dynamic_cast and RTTI in PlaneShift, and before we
stop doing so, we need to discuss the tradeoffs and agree on a course of
action as a team.

This is also not the time to be changing 59 files; please hold off on
such changes until after 0.5 is branched.  trunk is (unofficially) frozen.