Fix reimplementation of round() to be correct for negative values.
Although we'd all like to point and laugh at Microsoft for not supporting
cutting edge functionality like the ability to round numbers - a feature
added to the C language a mere 10 years ago - they may not deserve it.

It turns out that round() isn't part of the official C++ standard.

Instead, be disgusted at the fact that you have to pay $30 for a
downloadable copy of the C++ language specification.  Poppycock.