Revert change to show character stopped at sector crossing to others.
This reverts the change to  stop in the new sector and send a DR update
message to the server. That was causing the client to crash due to having
incomplete CS engine structures for the sector (they are being loaded in
a separate thread). This usually happened on teleport to a new sector
instead of crossing to a visible sector.
The problem with using the special "loading sector" is that the server
doesn't fully support it and sends a forced DR postion update to the client
thus causing a loop. The real issue is that the server thinks the player
is in the new location (and sends that info to other near by players) but
the client can't display the new sector until it is loaded.
So either the client needs to be better at not accessing the CS engine
structures being loaded or the server needs a better way of handling the
loading sector.